move data loading to infosheet content
